Move token price model to common (#95)
