+++ common/models/token_prices.py
@@ -1,0 +1,9 @@
+from sqlalchemy import Column, DateTime, Numeric, String
+
+from common.models import HemeraModel
+
+
+class TokenPrices(HemeraModel):
+    symbol = Column(String, primary_key=True)
+    timestamp = Column(DateTime, primary_key=True)
+    price = Column(Numeric)
